requirements.txt
efficientnet_pytorch==0.6.3
albumentations==0.5.2
numpy==1.18.1
opencv_python==4.2.0.34
matplotlib==3.3.3
torch==1.7.0
pandas==0.25.3
torchvision==0.8.1
Pillow==8.1.0
imutils==0.5.3
